FPS and mins not showing up on "Pipelines" widget

+++ packages/plugin-sdk/src/contracts/dashboard.ts
@@ -124,14 +124,14 @@ export const DASHBOARD_SCHEMA = /* GraphQL */ `
 
   type PipelineModelMins {
     model: String!
-    mins: Int!
+    mins: Float!
     sessions: Int!
     avgFps: Float!
   }
 
   type PipelineUsage {
     name: String!
-    mins: Int!
+    mins: Float!
     sessions: Int!
     avgFps: Float!
     color: String
